financial_chatbot/requirements.txt
google-generativeai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-google-genai==2.1.8
google-ai-generativelanguage==0.6.18
python-dotenv